+++ CHANGELOG.md
@@ -14,7 +14,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 - [#480](https://gitlab.com/dctx/rapidpass/rapidpass-api/-/issues/480) `2.0.3.alpha.1` Added register scanner device
-    endpoint.
+    endpoint via Keycloak.
 
 ## [2.0.2] - 2020-05-19
 ### Added